Have JIDs implement net.Addr
I'm not entirely sure this makes sense; on the one hand, they are
addresses, on the other, the net package only supports a few protocols
and all of it's Dial methods take a string of the network type; making
up a new one probably breaks the point of having the Addr interface. I
suspect this was a bad design choice in the net package, as it rather
weakens the contract. Or maybe that's not the point of having a function
to return the network? I'm really not sure.

This should be considered experimental and may be reverted.