I figured out why Bind should still be a function
We don't want end users to be able to screw themselves over by changing
out the default Bind implementation. If they want a custom
implementation, they have to make their own and pass it in using the
normal Dialer APIs, not just modify the default one and have all the
dialers and things that use it suddenly start behaving differently.