xmpp: only error on flush to closed stream
I'm not entirely sure if this behavior is correct, but it "feels right"
and it lets us pass the decoder (with its existing state) directly to
the features (so that trying to wrap it with xml.NewDecoder in parse
will return the original decoder). We should probably just change the
API to accept one to prevent anything else from being passed and
breaking the state on parse.