xmpp: cleanup contexts incrementally
We're not blocking when checking the old context (there's a default
branch in the select) so the blocking I thought we needed to avoid
doesn't exist. Oops.
This is a race, but it's not a big deal, the user always needs to set
new deadlines before the old one expires.