all: add and use new internal/attr package
This finally removes all exported identifiers from the internal package
so that it doesn't turn into a junk drawer of functionality and
everything contains a name that's useful, not just a useless identifier
like "internal" that doesn't tell you anything about what the identifier
does.

Update everything else to use the new package as well.