mux: simplify buffer state tracking
Previously we kept separate state for any buffered token readers and
jumped back and forth between the start of the stream (when handing them
off to handlers) and the last read start element. This patch simplifies
the code by creating a new buffered reader sharing the underlying buffer
for handlers so that the offset never has to be changed in the original
buffered reader.

Signed-off-by: Sam Whited <sam@samwhited.com>