.builds: let the DCO check work on master branch
Previously we compared the branch against whatever the current master
branch was to find new commits, so if you push to your own master branch
or we test a merged master there would never be any commits to check.
Instead, check against origin/master so that we're always comparing with
whatever we pulled down from upstream.

Signed-off-by: Sam Whited <sam@samwhited.com>