all: change handler to provide a decoder
This patch changes the definition of handlers from taking an
xmlstream.TokenReadEncoder to the new xmlstream.TokenDecodeEncoder.
This is a breaking API change.

I've changed the definition of Handlers several times in this past, and
was loath to change it again becaues it's a major breaking change that
will affect almost all users of this module, but I decided it was
necessary once again.

While working on new packages, I was writing something like the
following as the first line of every handler:

    d := xml.NewTokenDecoder(tokenReader)

I had also introduced bugs that broke this several times (eg. because
the first start token had already been consumed and the new Decoder
would encounter the end token and break). This led me to think that we
needed to provide a decoder internally so that we could already use it
to pop tokens and make sure its internal state tracking was up to date
with the underlying stream, and to reduce the amount of boilerplate code
users of the Handler types have to write.

A simpler argument for the change might just be one of consistency: we
were previously returning an Encoder for writing tokens, but only a
TokenReader for reading them. Why would we provide more functionality
for writing tokens than we would for reading them? The answer is likely
that we can construct a Decoder from the TokenReader but we can't
construct an Encoder from a TokenWriter with the current functionality
of encoding/xml, but it's still nice for us to just provide a symetric
API to the users.

This feels like the logical next evolution of the API. I am not sure why
it wasn't done the previous time when we changed to an
xmlstream.TokenReadEncoder, but it is done now and I don't see anywhere
else we could really go with this particular parameter so hopefully
things will be more stable as we continue to progress towards 1.0.

Signed-off-by: Sam Whited <sam@samwhited.com>