internal/marshal: add encoder wrapper
Adds the ability to wrap an xml.TokenWriter into an xmlstream.Encoder
that uses the other encoding functions from the Marshal package under
the hood.

It would be better if we had something like xml.NewTokenEncoder similar
to the NewTokenDecoder changes we made to the standard library to
facilitate this module, but changes to encoding/xml are hard to get
reviwed and this would be a much bigger change than the equivalent token
decoder changes, so for now we will continue to use the inefficient
workarounds from the marshal package.

Signed-off-by: Sam Whited <sam@samwhited.com>