roster: fix panic when closing errored roster iter
When requesting the roster we create an iterator. If an error occurs, we
defer reporting the error until the iterator is used and the internal
xmlstream.Iter that it holds will be nil. Unfortunately, this caused a
panic if we defer a call to iter.Close. Closing a nil iterator is a noop
and is expected behavior, so don't report an error and don't panic in
this case.

Signed-off-by: Sam Whited <sam@samwhited.com>