xmpp: provide access to tls.ConnectionState
When a session's underlying net.Conn gets wrapped, the underlying
tls.ConnectionState becomes inaccessible. Saving the connState and
adding a method on Session to access it will let us proxy it forward to
things that need it (eg. SASL) so that they don't break if we negotiate
stream features that wrap the conn.

See #45

Signed-off-by: Sam Whited <sam@samwhited.com>