xmpp: proxy ConnectionState through teeConn
When we use the TeeIn/TeeOut functionality our connections get wrapped
in a teeConn which would hide the connection state if the underlying
connection is a *tls.Conn (or anything else that exposes TLS connection
state). Proxying it through lets us access the connection state from all
the way up at the Session level (which means that SASL mechanisms can
get access to it).

See #45

Signed-off-by: Sam Whited <sam@samwhited.com>