docs: mention autosigning commits in CONTRIBUTING
This explains how to autosign commits in the CONTRIBUTING doc because I
found a lot of conflicting information about this when searching online.
It also moves the signing section up to just below the commit message
section since it's part of the commit message and it makes sense that
even if it's not part of that section, it is related and should be
grouped near it.

Signed-off-by: Sam Whited <sam@samwhited.com>