roster: remove workaround for encoding/xml bug
An issue existed in versions of Go prior to 1.14 that caused the decoder
to error if the underlying xml.TokenReader returned a token and io.EOF
at the same time.
Now that this is fixed in all supported versions of Go, remove our
workaround.

See https://golang.org/cl/130556

Fixes #29

Signed-off-by: Sam Whited <sam@samwhited.com>