mux: move options into their own file
This is a minor refactor that just oves options out into their own file.
While working on the disco package I decided to move them because its
integration with the mux package was going to add even more options and
they were getting a bit hard to find in the sam file as everything else.
Going ahead and moving them out keeps the diff smaller during that work
and makes it easier to see what has changed.

Signed-off-by: Sam Whited <sam@samwhited.com>