internal/integration: simplify use of ClientCert
This patch lets us promote cmd.ClientCert to a function and use it as a
tls.Config's GetClientCertificate function.
It also requires that we write out the client certs to disk first so
that if we need to debug later we have them and don't have to guess at
what was happening.

Signed-off-by: Sam Whited <sam@samwhited.com>