xmpp: remove invalid TODO comment
This was already fixed which caused me to misunderstand it when filing
an issue to remind myself to fix this. I thought it meant "remove the
cache" but it meant that the stream feature parse data used to be in the
cache as well and to remove that. Since we've already removed it the
comment and the issue can be closed.

Fixes #79

Signed-off-by: Sam Whited <sam@samwhited.com>