dial: fix fallback when one service returns "."
Previously if one service (eg. xmpp-client) returned "." and the other
(xmpps-client) did not, we wouldn't fall back and instead would return
an error. This patch fixes that to always use whichever service is not
explicitly set to ".". It also adds several integration tests against
the bad configurations at badxmpp.eu.

Signed-off-by: Sam Whited <sam@samwhited.com>