internal/marshal: better efficiency for some types
Previously if we wanted to encode an arbitrary type to a token writer we
were forced to encode it to a byte buffer then decode that buffer and
copy it to the writer. This was extremely inefficient. To help us fix
this for at least the types in this module, short circuit if the type is
any of the various interfaces that can be copied directly to the writer
and use their methods for doing so. This patch handles short circuiting
on xml.TokenReader, xmlstream.Marshaler, and xmlstream.WriterTo.

See #38
Fixes #73

Signed-off-by: Sam Whited <sam@samwhited.com>