ping: improve round trip test using ClientServer
Previously the test directly ran the get function and the handler
against a buffer. For a simple ping this is fine, but even for this it
requires a lot of boilerplate.  Using the new(ish) xmpptest.ClientServer
lets us excersize the code in a way that's closer to using an actual
server.

Updates #84

Signed-off-by: Sam Whited <sam@samwhited.com>