xtime: fix unmarshaling by changing the Time type
Previously unmarshaling was broken because of some pointer/copying
issues. This changes the type to always unmarshal from the correct
element and fixes the round trip test to detect this sort of issue in
the future.

The refactor for this test also fixes the other place where we wanted to
refactor the roundtrip test to use ClientServer.

Fixes #84

Signed-off-by: Sam Whited <sam@samwhited.com>