stream: add ErrorReader
Added a new xmlstream.Transformer that automatically unmarshals and
returns stream errors (but does not handle other stream namespaced
elements). This will be useful in places where we hand off stream
handling to the user (eg. calling a stream features Negotiate function)
and we want to handle any errors that arise during negotiation.

Signed-off-by: Sam Whited <sam@samwhited.com>