all: change stream feature Parse signature
This is a breaking change that makes the function signature of the Parse
function on StreamFeatures actually do what everyone is going to
immediately do with it anyways.

It also fixes the API to make it clear that we're passing a Decoder
which we had to do anyways because if we passed a wrapped TokenReader
and the user created a new decoder from that it wouldn't have had the
start element that we pass in popped from it and would error on reachign
its end element. This wasn't happening only because we were passing the
Decoder anyways and the NewTokenDecoder function would return the
original if the TokenReader was already a Decoder.

Signed-off-by: Sam Whited <sam@samwhited.com>