internal/integration/ejabberd: wait after stop
The "ejabberdctl stop" command was run asynchronously which was causing
a race condition that would sometimes cause the next tests to fail with
an error about ejabberd already being started.
Block until its completely stopped with a second call to the
"ejabberdctl stopped" command.

This doesn't appear to actually fix the issue (on FreeBSD at least), but
it should be done either way and will likely help.

Signed-off-by: Sam Whited <sam@samwhited.com>