component: only check stream ID if not an error
This fixes an error where the stream ID wasn't set on Prosody unless the
stream would be used. While I think Prosody is in the wrong here, their
behavior likely won't cause any problems so it doesn't hurt us to mimic
it for compatibility as long as we still maintain the check on streams
that will be used.

Signed-off-by: Sam Whited <sam@samwhited.com>