internal/integration: wait for shutdown of command
Always wait for commands to shutdown during Close so that any final log
messages are flushed before the tests exit.
This may hang forever if we haven't set the Shutdown option, but right
now all integration test packages set this so we don't have to worry
about it. If it ever becomes a problem in the future we can add a
default shutdown function that just sends SIGTERM.

Signed-off-by: Sam Whited <sam@samwhited.com>