internal/integration: fix tempfile and test naming
If the command specified a full path we would previously attempt to
create temp files that included the full path (including slashes) which
would fail. Similarly, tests would be named including the full path
(which was long and confusing looking). Fix this by always using only
the basename of the command.

Signed-off-by: Sam Whited <sam@samwhited.com>