xmpp: support stream features that use IQs
Previously we were passing the start token in as the data parameter for
server side features, but this made us have to awkwardly reconstruct it
to the stream on any feature that needed it. Instead have the Negotiator
put the start element back on the stream before handing it off to the
stream feature.

This also allows stream features that use IQs such as resource binding
and legacy in-band registration by first popping the IQ token and
matching based on the payload instead of the stanza start.
This allows the server side of resource binding to work at long last.

Fixes #97

Signed-off-by: Sam Whited <sam@samwhited.com>