xmpp: improve some error messages
A number of functions return stream errors due to an old API that was
never actually used in this library and a lot of copy/paste even though
no stream error was encountered. This patch fixes a handful of these
errors to return meaningful error messages that make debugging easier
and don't lead to confusion about stream errors being transmitted when
the other side did no such thing.

Signed-off-by: Sam Whited <sam@samwhited.com>