xmpp: always set namespace on stanzas
Attempt to guess at stanzas and set a namespace on them if unset. This
guarantees that the other side actually sees them as stanzas and the
user doesn't have to know what type of session we have and remember to
set a namespace every single time they send anything.

This won't change anything for the normal TCP binding (where there was a
default namespace and the empty namespace would just use that), but for
future bindings like WebSockets where there isn't a default namespace
this will make sure IQs still get the xmlns attribute written.

Signed-off-by: Sam Whited <sam@samwhited.com>