all: update negotiator API to handle stream info
Previously we were special casing this in our negotiator implementation,
meaning you couldn't actually implement the negotiator in another
package. This fixes the API so that it could in theory be moved to its
own package (if we could do so and avoid import loops somehow).
This also moves us a step closer to supporting virtual hosts by making
sure the input/output stream info is set before the negotiate call is
made.

See #61

Signed-off-by: Sam Whited <sam@samwhited.com>