xmpp: fix broken LocalAddr and RemoteAddr methods
These were deliberately broken by the changes to how we handle to/from
attributes. Fixing them is one of the last things that needs to be done
for handling virtual hosts as well because we can now get the current
local address in the feature negotiation callback from the stream header
and use it to determine what features to offer.

Fixes #105
Updates #61

Signed-off-by: Sam Whited <sam@samwhited.com>