internal/genfeature: new tool for disco features
This tool allows us to generate disco features for the many packages
that need them. While this is likely a one time change and a tool
doesn't reduce the workload much, there is a chance I'll want to change
what these types look like later so having a tool generate them makes it
easy to update the tool and then not have to redo all the work in each
individual package.

Signed-off-by: Sam Whited <sam@samwhited.com>