xmpp: always start a server goroutine
Previously it was up to the user to start the server goroutine and if
they did not methods that rely on it such as SendIQ would block forever.
This patch changes this by always setting a server with a default
handler and allowing the user to update the handler (and block until the
server goroutine finishes) with a call to `Serve'.

The behavior of multiple `Serve' calls is still undefined.

Signed-off-by: Sam Whited <sam@samwhited.com>