stanza: add Error method to Presence and Message
Like IQs replying with an error involves switching the to/from
attributes and making sure the type is set to "error" so it makes sense
to have the Error method on both Presence and Message too.

While doing this I noticed that each stanza marshaled its attribute
order differently. While it doesn't matter and this is not a guaranteed
part of the API, it would be nice if they were consistent so I also
switched those around and fixed a few minor documentation issues.

Finally, since we were adding a Presence and Message Error method with a
test I also added a test belatedly for the IQ Error method.

Signed-off-by: Sam Whited <sam@samwhited.com>