xmpp: use function for Web Socket negotiation
Previously this was handled by a WebSocket bool on the StreamConfig
struct, but doing it this way means that with the new API stream
features could switch back and forth between the websocket profile and
the normal TLS profile of stream negotiation. This doesn't make much
sense. Having them as two separate functions makes more sense from the
users perspective and makes it easier to discover and document the
behavior. It's also a stepping stone for eventually moving the websocket
negotiator to the websocket package.

Signed-off-by: Sam Whited <sam@samwhited.com>