xmpp: return start token for iter IQ payloads
Previously the start token was discarded when calling IterIQ or
IterIQElement. However, we may need attributes from it, making IterIQ
impossible to use for some cases where we still might want an easy way
to iterate. Returning the start element fixes this and avoids a lot of
boilerplate re-implementing iter IQ for every place we might want to do
this.

Signed-off-by: Sam Whited <sam@samwhited.com>