design: update service discovery proposal
The previous plan to include a registry was a rather naive
implementation that would have required that the multiplexer and the
registry both have features registered against them, resulting in a lot
of boilerplate and potential for advertising features that don't exist
or failing to advertise features that do exist because we forgot one or
the other registration.

A new plan was created that uses the multiplexer as the registery so
that handlers regstered against the mux that also advertise features
can automatically respond to disco info requests.
The new plan has the added benefit of working with nested multiplexers
without further modification.

Signed-off-by: Sam Whited <sam@samwhited.com>