docs: remove references to the mailing list
Since we're likely not going to continue using SourceHut in the long
term, remove mentions of it from the contributing document until such a
time as we decide where our long-term home will be.

As much as I like the patch based workflow, SourceHut doesn't do a very
good job of making it accessible to new users and I don't like the idea
of asking contributors to setup a new workflow that requires more than
just git(1) and the forge we're pushing to.

Maybe one day something will integrate a patch based workflow into a
codeforge well and we'll be able to add it back.

Signed-off-by: Sam Whited <sam@samwhited.com>