internal/discover: remove extra fallback addresses
Fallback currently breaks if, eg. port 80 has HTTP on it and we connect
because we're just testing whether we can establish TLS, not whether we
can actually get a response that looks like XMPP.

Until we can figure out a better way to probe ports that doesn't involve
violating the separation we keep between layers in the stack, just
remove the fallbacks.

Signed-off-by: Sam Whited <sam@samwhited.com>