history: add support for tracked queries
Previously the history package contained minimal functionality for
sending queries and receiving responses asynchronously.
This patch adds support for tracking queries and iterating over their
results in the Go style of writing async code that, for all intents and
purposes, appears synchronous.

Like IQ tracking, this is accomplished by adding Fetch methods to the
handler that are similar to the Fetch functions except that they store
their results on the handler itself.
When a message is incoming the ID is checked and if it is a part of a
tracked query it is sent over a channel to the iterator, otherwise it is
handled like any other untracked messages.
If an iterator is closed it unregisters itself and any subsequent
messages that are received are handled just like any other untracked
message.

Signed-off-by: Sam Whited <sam@samwhited.com>