history: support paging before and after
Previously there was no way to support the "before" and "after" queries
that are a part of Result Set Management.
This was because they are more or less redundant with after-id and
before-id as far as I can tell.

However, since it turns out that not much supports mam:2#extended (and
thus before-id and after-id queries) that the older before/after filters
can be used alongside start/end time based queries to effectively
emulate before-id and after-id.

Though it still feels like they're more or less the same thing and I
don't fully understand the difference, this adds support fore
before/after queries.

Signed-off-by: Sam Whited <sam@samwhited.com>