mux: ignore namespace when matching stanzas
Previously we checked that stanzas were actually stanzas (they had the
namespace jabber:client or jabber:server), however, this check is not
technically correct because there is also the component namespace.
For now, remove the check and treat all elements named "message", "iq",
or "presence" as stanzas. In the future we should improve this to allow
the negotiator to set the stanza namespace on the session and only check
for the specific correct namespace which would be more correct than this
solution or the previous way.

Signed-off-by: genofire <geno+dev@fireorbit.de>