styling: remove style from tokens
Previously the current style was present on both tokens and the decoder,
leading to unnecessary duplication. Removing the style from individual
tokens makes the token stream re-orderable and reflects the fact that
styles aren't a property of individual tokens because changes to other
tokens earlier (or later) in the stream can affect tokens in the middle.

Fixes #129

Signed-off-by: Sam Whited <sam@samwhited.com>