styling: remove style from tokens
Previously the current style was present on both tokens and the decoder,
leading to unnecessary duplication. Removing the style from individual
tokens reflects the fact that styles aren't a property of individual
tokens because Message Styling is a context-sensative language and
changes to other tokens earlier (or later) in the stream can affect the
style of other tokens around them.

Fixes #129

Signed-off-by: Sam Whited <sam@samwhited.com>