make: add rule to generate dependency graph
This adds a rule to generate a graphic representation of the dependency
graph using graphviz.
This was previously added in fad579be6a84 but after the Makefile was
removed in 804efe9cce09 and then re-created in a1400b09ec02 for the
CONTRIBUTORS rule the graph rule never got added back.

Signed-off-by: Sam Whited <sam@samwhited.com>