xmpp: write error IQ responses correctly
While working on IBB tests the other side of the connection sent us a
payload we don't support, resulting in an error. However, I noticed the
error was showing up in the logs twice: once being sent and once being
received. It turns out we have been sending errors to ourselves instead
of the remote entity. This patch fixes that.

Signed-off-by: Sam Whited <sam@samwhited.com>