ibb: simplify data handling
We were previously reading from either a buffer or directly from a pipe
which was sometimes being copied to the buffer if no call to Read was
outstanding. This resulted in deadlocks and very complicated code.
Instead, simply lock the buffer before writing or reading from it and
block waiting for new data on a signal (instead of using a pipe).

Signed-off-by: Sam Whited <sam@samwhited.com>