ibb: fix data race when accepting connection
Incoming open calls passed to a listener currently happen in a different
goroutine than the goroutine creating and destroying the listeners,
resulting in a race condition where an open call might come in, then the
listener is destroyed, then we continue processing the call using a
listener that shouldn't exist.
Adding a mutex around the listener map solves this issue.

Signed-off-by: Sam Whited <sam@samwhited.com>