xmpp: minimize lock contention on handlers
Handlers take out a read and a write lock on the stream. These locks are
not released until the handler returns, resulting in a great deal of
lock contention even if the handler never ends up reading or writing.

This patch minimizes lock contention in two ways:

For the reader we limit the handler to reading a single top level XML
element. This means that as soon as reading the element is complete we
can go ahead and release the lock. In the worst case (the handler
doesn't completely read the element) we have the same issue. In the best
(and likely most common case), the handler decodes the entire element
immediately and the lock can be released before the main body of the
handler begins.

For writes the solution is similar: we defer taking the lock out until
the first call to EncodeToken. This way, if we never write anything (ie.
in the common case where a positive IQ response will be sent which
happens by default if one isn't already written), the lock is never
acquired at all!

Fixes #83

Signed-off-by: Sam Whited <sam@samwhited.com>