ibb: implement max buffer size
Add the ability to restrict the buffer size so that it doesn't just grow
forever. When the buffer is full, send back a resource-constraint error
of type 'wait' to signal to the other end that they can retry the
transfer shortly (hopefully after we've read some data and shrunk the
buffer).

Fixes #239

Signed-off-by: Sam Whited <sam@samwhited.com>