ibb: wait for all data in slixmpp integration test
Previously there was a race condition where the test would end before
the connection was closed, causing incomplete data to be transmitted.
The timing happened to be consistent enough that this was never observed
until we made some minor changes to an unrelated part of the package
that caused a context switch and allowed the race condition to affect
the results of the test.

Signed-off-by: Sam Whited <sam@samwhited.com>