websocket: fix incorrect TLS host verification
Previously after fetching a websocket connection endpoint from a DNS TXT
record we were verifying the certificate against the name fetched from
the record. This is incorrect and results in the ability to MITM the
connection because if DNS is spoofed and the WSS endpoint is changed to
evil.example.net we would verify that evil.example.net presented a valid
certificate for itself, but not for good.example.net (or whatever the
original address was expecting). Instead, verify the original address.

This issue has been assigned CVE ID CVE-2022-24968.
For more information see: https://mellium.im/cve/cve-2022-24968/

Fixes #259

Signed-off-by: Sam Whited <sam@samwhited.com>