xmpp: add informational stream features API
Previously all stream features would be negotiated if they were
supported by both the client and the server.
However, some stream features such as the one specified in XEP-0115:
Entity Capabilities [1] for advertising the servers capabilities are not
meant to be negotiated, they are informational only.
To allow this, skip negotiation of any feature that does not actually
have a Negotiate function, making informational stream features as
simple as not setting this function to a non-nil value.

[1]: https://xmpp.org/extensions/xep-0115.html#stream

Fixes #263

Signed-off-by: Sam Whited <sam@samwhited.com>