component: ensure stream namespace is set
This patch ensures that we're reading the <stream> start tokens through
the session's decoder (not a new one that we create and then discard) so
that the default namespace is correctly set.
Prior to this patch no stream namespace was set and therefore anything
that matched on it (eg. the mux package) would expect a namespace of
"jabber:component:accept" but would see a namespace of "" and ignore it.

Fixes #293

Signed-off-by: Sam Whited <sam@samwhited.com>