ibr2: deprecate old package
The ibr2 package implements an old version of the extensible in-band
registration protocol and we currently don't have plans to update it to
the latest version.

This patch adds a deprecation warning so that the package can be removed
in the next release after deprecation is complete.

The test file was removed due to a bug in our linter that caused it to
complain about the ibr2_test package importing a deprecated package (but
of course it does).

Fixes #295

Signed-off-by: Sam Whited <sam@samwhited.com>